refactor: update utils files
