+++ lib/src/widget/built_in/minimal/minimal.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
 import 'package:flutter/material.dart';
-import 'package:toastification/src/core/context_ext.dart';
+import 'package:toastification/src/utils/toast_theme_utils.dart';
 import 'package:toastification/src/core/style/toastification_theme.dart';
 import 'package:toastification/src/widget/built_in/built_in.dart';
 import 'package:toastification/src/widget/built_in/widget/close_button.dart';
